Rearrange the RocketsPage
