+++ src/__test__/RocketsItem.test.js
@@ -1,5 +1,4 @@
 import React from 'react';
-// eslint-disable-next-line import/no-extraneous-dependencies
 import renderer from 'react-test-renderer';
 import { Provider } from 'react-redux';
 import store from '../redux/store';
