+++ src/components/rockets/Rockets.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { useSelector } from 'react-redux';
+import classes from './Rockets.module.css';
+import RocketsItem from './RocketsItem';
+
+const Rockets = () => {
+  const rockets = useSelector((state) => state.rockets.rockets);
+
+  return (
+    <section className={classes.rocket__list}>
+      {rockets.map((item) => (
+        <RocketsItem item={item} key={item.rocket_id} />
+      ))}
+    </section>
+  );
+};
+
+export default Rockets;
